refactor(st-09098): share plan-execute model parsing
